Infoitems adde
Infoitems are now parsed correctly from the plist. If the link to an
item is an external link, it will open in the webbrowser(IE) if not and
it 's relative, a webcontrol is used. If the content has subcontent,
then app navigates to a view of these contents and action from there are
the same as in the first case