Set the tint color of the feed refresh control
This commit set the tint color of the refresh control in the feed collection
view to white. White matches the rest of the interface better than the current
gray and is more 'legible' on the dark blue background. It also moves setting
the alwaysBounceVertical property to true from code to Interface Builder.