[WIP] Re-organize the repo to make automatic deployment easier
- The server folder contains all code that should be pushed to the
  server. Currently, the plan is to still use git pull to get changes,
  which means we need the full repo. However, in the future, Travis
  could rsync the folder instead.
- The server/assistant folder contains the code for the Google Assistant.
  This changes requires a change in the ansible-config!
- server/scripts contains all scripts necessary to run stuff. It contains
  the deployment scripts, and the scripts needed to run the scrapers and
  other data producers.
- server/api is a git-ignored folder that contains the outputted data
  from the api stuff. This folder is currently not tracked, but should
  not be deleted (see TODO in deployment.sh)
- server/src contains the Python scripts and other static data for the
  API.
- server/tests contains tests run by Travis before deployment
- server/website contains the website. Since this is currently just a
  raw website, the folder in the public part is symlinked.

Note that the folder structure/naming is not yet fixed and may change
if problems or ambiguities arise.

The goal is to have Travis automatically deploy the repo to King and
have it set up things if a commit is made on the master branch. Currently,
(in my opinion), the process to get something deployed is really slow
and error prone: if someone commits something/merges a PR, they not only
have to find someone with access to King, but who also knows how things
are set up. I believe automating all those tasks is the best documentation;
in doubt, one can always view the scripts to see what happens.