Add UGent news scraper
We used to use the Plone rss2json format directly. With an update of the
Plone CMS, this feature has been removed. DICT has been so kind to
re-enable RSS for us, so we now manually convert it to JSON every day.
This is a lot easier for the apps, since they don't need an RSS or ATOM
parser.